+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Added mask in function for aperture extraction to mask out bad data points.
+
 ### Added
 
 - Added function to perform background subtraction.
@@ -21,5 +25,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Aperture photometry function in `extract.py`.
+- Aperture extraction function in `extract.py`.
 - CHANGELOG.md created